refactor: ディレクトリ名を修正
